pyproject.toml
[project]
name = "zenoviz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "boto3>=1.42.90",
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.13.2",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "anyio>=4.13.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
